requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
distro==1.9.0
etelemetry==0.3.1
fastapi==0.115.3
filelock==3.16.1
fitz==0.0.1.dev2
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
isodate==0.6.1
jiter==0.6.1
joblib==1.4.2
loguru==0.7.2
looseversion==1.3.0
lxml==5.3.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.8.6
nltk==3.9.1
numpy==2.1.2
openai==1.52.2
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
pgvector==0.3.5
pillow==11.0.0
prov==2.0.1
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
pyparsing==3.2.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
rdflib==6.3.2
regex==2024.9.11
requests==2.32.3
scipy==1.14.1
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
starlette==0.41.0
tqdm==4.66.5
traits==6.3.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
